Repair All Done

+++ database/migrations/2018_12_19_133507_create_mcb_transaction_table.php
@@ -24,8 +24,8 @@ class CreateMcbTransactionTable extends Migration
             $table->integer('mcb_id')->unsigned();
             $table->integer('block_id')->unsigned();
             $table->integer('category_mcb_id')->unsigned();
-            $table->tinyInteger('stream_reff');
-            $table->tinyInteger('voltage_reff');
+            $table->double('stream_reff');
+            $table->double('voltage_reff');
             $table->foreign('mcb_id')
                 ->references('id')
                 ->on('mcb')
